[#17094639, #21000115] Add URLs to partner sites URL conf for RSS fedds and at the same time align those paths more closely to the paths of corresponding web pages. Change the name attribute of some urls to align them both between Akvo RSR and partner sites and also to a more homogenous naming scheme.
For partner sites there are two RSS feed URL paths:

/rss/project/NNN/updates/ where NNN is the ID of a project
Returns all updates of that project

/rss/organisation/NNN/updates/ where NNN is the ID of an organisation
Returns all updates for all projects that the organisation is partner to

Discussion
==========
A separate URL conf file for partner sites is necessary given that the paths on partner sites and akvo RSR are
different. However I think the paths should be as closely related as possible since we will probably remove the /rsr
part of the path structure on Akvo RSR in the future.

Another reason for aligning the paths as much as possible became apparent when the RSS feeds were to be used for
partner sites. The name attributes for those URLs _have_ to be the same in both URL confs to be able to call them
via object.get_absolute_url() which is the recommended behaviour for the Django RSS framework. And if they are named
the same the paths should be the same or as close to the same as possible to minimize confusion.

This commit only begins the job of path and name alignment.