Fix csrf error in forms when using iframes on PS
Some partner are using iframes instead of DNS to make the partner site
live under their domain. When partners use an iframe to host the
partner site and the user have not earlier signed in from the raw
partner site *.akvoapp.org domain the correct csrf cookie will not be
set. Hence we need to make sure that the user request the page with the
form with a GET on the *.akvoapp.org domain. This fix will make sure
the "sign in", "sign out" and the "add update" links breaks out from
the iframe by adding a _top target. This change does not add any
redirect on logic to GET views with forms. Hopefully we will not
need to add that complexity to solve this issue.

[Fixes #29907575]