Update permissions for RSR editors and RSR managers
The permissions for the RSR editors and RSR managers groups were missing some
models resulting in not all inline forms being visible in the admin when
upgrading to Django 1.4.
This is now fixed and the permissions for RSR managers were extended with the
set of permissions for RSR editors removing the need to belong to both groups.

[Fix #29704531]