Make partner sites use i18n urls
All pages on partner sites are now localized, that implies that if you
request "/" you will end up to "/<locale>/" where <locale> is the
default_language (en) field on the partner sites model. There is a new
middleware that replaces the default Django LocaleMiddleware to support
the urlconf swapping that our partner sites middleware does.

There is a new language chooser which sets the window.location to the
current page with the chosen language prepended.

[#29625905], [#29702725]. #6