Location model refactoring
Change the Location model to use foreign keys instead of generic relations.
This is accomplished by creating an abstract base class, BaseLocation, that
defines all the location data fields except what object it is related to.
BaseLocation is then subclassed into OrganisationLocation and ProjectLocation
that have foreign keys to their respective models. By adding a primary_location
FK field on Organisation and Project, which replaces the model methods with
the same name most of the functionality in RSR just continues to work after
the refactoring.

Change the admin to use the new Location models. This includes adding form
validation checking that there is exactly one primary location if a project or
organisation relates to any locations.

Add the South migrations needed for the Location refactoring.

[#27]