Fix runtime error when deleting a project
Cascading deletes of invoices when a project is deleted may generate
runtime errors when signals update_project_budget() and
update_project_funding() are called by the invoice to be deleted.
Fix this by adding a try-except in the signals to ignore the attempted
update when the project doesn't exist any more.

[Fix # 59]