[#72] Add generic method for including related data in resources
resources.py
Add two new querystring parameters to the API, "depth" and "full".

"depth" is an integer and is used to indicate that related resources up
to depth relations removed from the main resource are to be included as
data rather than as URIs in the respose.

"full" is boolean. If full=True when "depth" is present all instances of
the same object are returned as full data, otherwise only the first
intance returned in the response is data, others are URIs.

Add ConditionalFullResource from which all resources inherit changed
get_list() and get_detail() methods. This is to add an object,
BundleRelatedDataInfo, to the bundle that holds meta data for use with
the"depth" qurystring parameter.

Add filtering to all resources. All related fields are filterable and a
selection of other fields.

Add fields.py to api
Add ConditionalFullFieldMixin, ConditionalFullToOneField,
ConditionalFullToManyField classes to fields.py to override the
dehydrate and dehydrate_related methods on ToOne and ToMany fields.
Also used with "depth" parameter