Use exception handling to ignore existing indices
Migration 0025 adds a number of indices to the rsr tables. For some
reason some of those indices already exist and an exception is thrown
when running the migration on test.akvo.org. To avoid this each index
creation call is wrapped in a try: except

Remove unused field from InvoiceResource.fields attribute