[Fix #92, fix #96] Move IATI ID to project funding page, remove dupes in project partner list
The IATI activity ID was moved to the project funding page, but it was
forgotten on partner site. Fixed.

The list of partner organisations on the project main page of both RSR
and partner sites shows the same organisation multiple times if it has
multiple partnership roles in the project. This is fixed by using the
Project.partners many-to-many field rather than the
Partnership.partnerships related name.