[Fix #91] Akvo administrators can view unpublished projects
Rename decorators.project_page() to
decorators.project_viewing_permissions.

Refactor decorators.project_viewing_permissions to not use the request
object for context passing.

Check if a user has the "rsr.change_project" permission and if so set
privileged_use to True. privileged_user is used to determine if the user
is allowed to view the project when it is unpublished, and also allows
the user to add updates.

Change the wording of the project draft view message.

Comment out unused context in views.projectmain()