[#121] Add conditional authentication
Add ConditionalApiKeyAuthentication so that authentication can be
required only for certain methods on a resource, e.g. for POST but not
for GET calls.

The methods requiring authentication are set when instantiating the
ConditionalApiKeyAuthentication object, e.g.
authentication = ConditionalApiKeyAuthentication(('POST', 'DELETE')).
If no params are specified authenticaton is required for all methods.