[#121] Add custom_get_or_create_country to rsr/utils.py
custom_get_or_create_country creates a new Country object for a given
iso_code if none exists in the database, or returns an exising object.

This is used in the admin when a user needs to create a new Country and
is also used in the API to create Country objects when POSTing data.