[#145] Refactor Project admin form
Add custom template logic to project/change_form.html that lays out the
form with the inline forms at "logical" places rather than at the end
as is the default.

The technique for getting the inline forms in the right places is to add
them at the right count in the loop ever the fieldsets. This is very
fragile, so any changes to the order of the fieldsets has to be matched
by changes in the admin template.