[#149] Add script that generates python tuples variables from IATI code lists
Add a script that creates tuples generated from the IATI code lists.
Each tuple is named after the code list but prefixed with IATI_LIST_ and
the words in the code list name are separated by _ and uppercased.

The tuples generated can be used in django model fields for choices
lists. In cases where there are more than two items for each code in the
IATI list additional munging will be needed to create a tuple of
two-tuples that django choices needs.

Add rsr/iati_code_lists.py with a tuple generated from the IATI
Organisation Type code list.