[#138] Remove span with notranslate attribute from title tag
Since we don't want Google to translate the organisation names I tried
adding a span tag with a notranslate attribute around the org in the
title. This doesn't work however since you cannot have markup in the
title.

I can't see how to solve this without resorting to some javascript
manipulation of the title.