[#121] Add IATIPartnershipResource.hydrate_organisation()
hydrate_organisation() tries to create the foreign key to the partner
organisation by first trying to find an exising org and failing that it
tries to create a new Organisation and link to that.

This way the import auto-generates new Organisations from the XML data,
provided that it is consitent and that all participating-orgs have at
least a name, either an IATI ID or an interal-ref ID from the
reporting-org and a type.

I'm not sure if this is the right way going forward to import
Organisations, but since there is currently so little information for
them it works for now.