[#121] Extend functionality of IATIPartnershipResource
The IATIPartnershipResource can now handle the creation of Partnership
objects and the linking of Organisation objects either creating the
Organisations on the fly of linking to existing ones.

There is quite a lot of custom code for Cordaid right now that should
be extracted from the general IATI resources (if possible).