[#121] Fix borked migrations
Replace migration 0032 to change empty string values in
Organisation.iati_org_id to NULL values to prepare for the unique
constraint on the field.

Add schema migration 0033 to set the unique constraint on
Organisation.iati_org_id and to create the InternalOrganisationID model.

Add data migration 0034 to add custom data to Cordaid and DGIS
organisation objects