[#121] Write importing script for images to the Cordaid IATI XML projects
The script creates a map between the photo-id and internal-org-id
attributes on the iati-activity tag in the XML. This enabels us to
identify the images, that are named as the photo-id, to be linked to
the correct project which we can find in the DB via the
internal-project-id.

The script loops over all images in the designated folder and tries to
find a matching project. If there is a match the image is saved to the
Project.current_image field.

Fix import in rsr/utils.py