[#207] Add notes field to PartnerSite model
Add PartnerSite.notes field to models.py

Add PartnerSiteAdmin.restricted_fieldsets leaving out the notes field
and add methods to PartnerSiteAdmin to choose which fieldsets to use
based on user permissions for the PartnerSite model.

Add PartnerSiteAdmin.get_list_display() to only show the notes field in
the list view for users with privileges as above.

Add migration for the PartnerSite.notes field.