[#214] Find unicode gremlins in API XML
Add script that calls the project, organisation and project_update
resources details, once for each object, requesting XML. If a call fails
it may be because there are unicode characters that aren't allowed in
XML, so we call the same resource requesting JSON format.
If this works it's probably because of one or more XML/unicode
"gremlins". This is noted in the script output.

Refactor cordaid_importer, moving the Requester helper class to its own
module to be used in xml_char_check.py