[#260] Fix link to RSR registering on partner sites
Since you have to register on RSR we need a link back to the
registration page in RSR on the partner sites sign-in page.

Also fix the link to Drupal content on RSR from same page. Note that
this is a hard coded link that'll probabl need fixing when the new
akvo.org site goes live.