[#224] Fix non-integer indicator values
Some indicator values from Cordaid are percentage floats, e.g. 6.4. The
Benchmark.value field is an integer field so these values have to be
rounded.

Fix by adding IATIBenchmarkResource.hydrate_value. Note that this method
sets value to 0 on all errors, if an exception occurs this should be
logged for traceability.