[#343] Fix map pins for API driven maps
The javascript that loads the maps using the API use a hard coded path
to the media root, '/rsr/media/' that doesn't work after the move of
RSR to rsr.akvo.org.

Fix by replacing the path with settings.MEDIA_URL as a context variable.

This also requires some tweaking of URL building to the API in
akvo-maps.js.