[#206] Change API to remove duplicates by default
When querying the API using the depth paramter there are cases where
you can get multiple copies of the same related object, when that
object is related to more than one object of the "root" kind.

Change this so duplicates are filtered out by default. If for some
reason the dupicates are needed use distinct=False in the query string
of the request.