[#438] Links in updates are urlized
If a link is present in the markdown update, but not marked
as a proper link we use the urlize template tag to create a
link from the url. This link will show the full url as the link

We do not urlize texts that are limited in chars. Examples of this
is the project list and the small preview on the main project page.
The motivation for this is that it might be unsafe to deal with.