[#451] Add org_data_from_xml()
org_data_from_xml() is a more general way to get from the etree
representation of the data to the dictionary needed in the Django ORM.

The core idea is to use an initial dict that holds the Organisation
field names as keys and the xpath expressions for getting the data from
the etree as values. After running the xpath part using calls to
data_from_xpaths() post processing of fields that need to be transformed
from the raw data is made.