[#419] Fix link in updates edge case
When an updates only consisted of a link that would
not get urlized. We have to do some escape dancing to
make sure urls are urlized by Django & that markdown
works properly. And at the same time be safe and not
allow things like <script>alert('unsafe');</script>
to be evaluated.