[#309] Point registering on Pages to step 2
When registering while at a Pages site the registration should default
to be with the organisation of the site.

Change the link to registration to reflect this, going directly to
step two.

Add a second link pointing to step one of registraion if the user wants
to register with another organisation.

Also set target on those links to "_blank" to open a new window/tab for
the registration since we can't redirect back correctly.