[#252] Add check for akvo name space
A number of times when running the IATI XML import the namespace
definition for the akvo name space has been incorrect. This leads to
errors of course, but it is very hard to trace the origin of the
probelm.

Fix by adding the AKVO_NS setting and add an assertion at the start of
the upload to make sure the name space is correct.