[#673] Fix filter drop-downs for Pages site using keywords
When a Pages site uses keywords to populate performace issues preclude
filtering the datasets used to populate the filter drop downs. This fix
reverts so that the filters are populated using the full datasets when
the site is keyword driven.

To enable this fix the middleware now adds the current partner_site to the
request so it can be determined if the site is keyword driven or not.