[#632] Enable the adding of Partnership objects
To be able to add Partnerships to Projects a way to do lookup of the
Organisation linked to in the Partnership is needed.

This solution does lookup via the API using the Organisation.iati_org_id
field primarily and if that does not exists lookup is done using
Organisation.name.

Add RSRModelInstance.partnership__organisation method.

Add more docstrings and comments.

Several small code fixes.