[#680] Make sure Project.primary_location is none for updated projects
When updating Cordaid projects the IATIProjectResource.obj_update
method deletes all related objects so they can be re-created. This leads
to Project.primary_location pointing to a deleted object so we have to
set it to None since QuerySet.delete() does not call Model.delete() for
the individual objects. (For reasons not understood a solution iterating
over the ProjectLocation objects and calling delete() didn't work.)