[#632] Add ValidXML field mappings
In RSR we use ValidXMLCharField and ValidXMLTextField in stead of the
vanilla Django Char and Text fields. DRF maps Django fields to its own
fields, but uses the fallback ModelField when it finds the RSR fields.

Fix by hooking into __init__ and adding mappings for the RSR fields, and
using the NonNullCharField to ensure that emtpy text fields return u''
rather than None.