[#632] Fix NULL URLField
The URLField suffers from the same problem as Char and Text fields, i.e.
if they are missing from the POST they are set to None by the serializer,
resulting in protests from the DB when saving.

Fix by creating NonNullURLField as a mixin class inheriting from
NonNullCharField and URLField.