[#632] Add simple write permissions
Add rsr_rest_user_<model_name> permissions for all rsr models. This
requires the running of manage.py syncdb --all to get into the DB.

Add IsRestUserOrReadOnly and set BaseRSRViewSet.permission_classes to use
it. This allows users with the rsr_rest_user_<model_name> permission to
use the "write" methods of the API (POST, PUT, PATCH and DELETE).

Suggested use is to create a Group with the rsr_rest_user_<model_name>
permissions and assigin users that are allowed to write to this group.