[#409] Updated TastyPie API with new user model
The new user model is incorporated in the TastyPie API. The UserProfileResource has been deleted, so we have to check whether no one uses it. Other than that, the API remains similar to the current version. However, a user can also have multiple organisations right now. This is mostly visible in the project_update_extra resource, where all organisations of a user are now displayed.