[#745] Add exclude keyword option to Akvo Pages
On Akvo Pages, a boolean field 'exclude_keywords' has been added. When set to False (default), the keywords act as usual. However, when set to True, projects with the indicated keywords will be excluded from the project list on the Akvo Page.

E.g. For the YEP Water page, we want to exclude the internal Akvo projects. By setting this option to True and indicating a new keyword ('akvo internal') on both the Akvo Page and the internal Akvo project(s), the project will not be shown on YEP's page.