[#733] Initial templates for RSR v3
Most of the old RSR templates have been placed in the templates/old folder. Instead, a new base template, header and footer templates (in /navigation), sign-in template, registration templates (in /registration) and a MyRSR template (in /myrsr) have been added.

In addition, Bootstrap has been used to create these templates. jQuery is used as well, which we might not want to use, so we still have to take a look at that.