[#783] Extended the User model with Employments
The User model has been extended with Employments, a new through relationship to the Organisation model. The fields 'is_approved', 'country' and 'job_title' have been added to the Employment model, as well as the 'avatar' field on the User model.

We still have to add the Employment model to the admin, and we'll have to adjust the form for requesting to join an organisation.