[#632] Add location_target to OrganisationLocationSerializer
Previously the location_target field was hidden in
OrganisationLocationSerializer to avoid emitting the organisation data
twice in the project_update_extra resource, since it includes (via
several intermediate serializers) the
OrganisationLocationExtraSerializer which in turn has a depth=1
resulting in the full location_target (i.e. the organisation) being
rendered inline to the location!

However for the Swagger API documentation to display a reasonably
accurate picture of the relationship between loactions and their
tagets/parents the location_target field should be included.

The solution is to specify location_target as a vanilla serializer.Field
and use the source attribute to retrieve the corect data (in this case
the ID of the Organisation).