[#797] Change password through API
The REST API has been updated with a new view on the usel model: /rest/v1/user/{pk}/change_password. A POST request to this view with the old_password, new_password1 and new_password2 will check if the credentials are valid and will update the user password.

Also, the form and view in MyRSR for updating a user password has been updated, now using the API through jQuery.