[#822 #824] Updated 'delete user' to 'remove organisation link of user'
For #822: Renamed the react-modal.jsx to react-user-management.jsx.

For #824: Remove the 'delete user' functionality and made it possible to remove an organisation link of user. Since we use the API to get data, the loading of the table can take a couple of seconds. We probably will need to render the data directly in the HTML instead to reduce loading times.