[#896] Add project_extra API endpoint
The requirements of Up are easiest to accomodate with a custom endpoint
for projects with some additional data.

This is solved by adding the project_extra endpoint. It provides for
limited filtering as described in a previous commit and also includes
the full data of the location objects associated with the project as well
as some of the data from the partnership objects.