[#848] On listing pagees map is now in sync
On the listing pages the map now reflects what in the table. The way
maps are loaded is reworked a bit. Now a coll_map template tag will
generate one div item with a random id & one script tag with a js
object that is referenceable via the same random id. At the js block
tag the google maps js library is loaded. In the common v3 library
there is a function that walks the elements of class rsr_map and
populates the maps with data from the js object.

Also fixed an issue realted to anonymous user and calling our user
class objects.