[#1617] Enabled Google translation bar option on pages
The translation bar can now be enabled per RSR page basis. This includes a
crude CSS tweak to make the translation bar and the RSR navbar happy. Could
not find a way to support default langauges so this option is disabled in
the admin.

Since we want to implement default lanaguage option traces is left in the
defaultLocaleMiddleware, but it's broken at the moment. This since it's
difficult to know if a path is an i18n one or not.