[Fixes #1616] Adds keyword logos to Akvo Pages.
If a project is tagged with a keyword that has a logo and the
'show keyword logos' is enabled on a Akvo Page 2 keyword logos
can be seen on the project main page. This option takes the
keyword logos from normal non Pages RSR to Pages but as opt in.