[#1793] Fixed + partners link in project directory
Cleaned up the more template tag to a bit since it was initially broken
on the project list page. Removing the two data attributes on the a tag
fixed that. Also made use of plurialized for the partner/s label.
Removed unused Tooltip Jquery since we got a general one in rsr.js which
is loaded on all pages.