[#1902] Fix migration 0040
Previous fix to the migration won't work since the "migration version"
of Project doesn't include custom methods.

Fix by importing the "real" Project model. I don't think this is
problem here since the migration is a pure data migration that has to
run in the context of the current schema.