[#1828] ICCO description fields
Rewrite the Description class, abstracting out the two general methods
of finding description fields matching the Project narrative fields.
Also prepare the class to be subclassed for special behaviour.

Add ICCODescriptions, a subclass that implements special rules for
importing IATI XML from ICCO, using a scheme where we look for prefixes
in the texts to match the Project fields to.

Also a few bugfixes found when testing against ICCO XML.