[#1895] Fix migration 0045
The data migration 0045 won't work when pk=1 is included when creating
the first object. At least on my vagrant box I get a unique constraint
error, so I'm removing the setting of the pk.

This means it's imperative to run migration 0044 and 0045 at the same
time to guarantee that the RSR validation set gets pk=1.