[#1980] Refactor IatiImportJob.run()
IatiImportJob.run() holds too much state and there are fail states not
taken care of.

Refactor so that all methods called to check if the import can be run
return True if all's well.

If a check fails, the method responsible logs both an error message
explaining the failure as well as a status change log setting the
status to cancelled.