[#1980] Fix scheduling isseus in IATI import
Fix setting of next execution time for scheduled imports that didn't
work correctly.

Fix problem with CordaidZipIatiImportJob resulting from the
refactoring of IatiImportJob in a previous commit.

Add debugging code that logs entry and exit of select methods. This is
a WIP to try to trace the performace problems the import seems to have,
and was used to get the scheduling right.

BUG: I think there is a bug remaing in CordaidZipIatiImportJob, see the
TODO in parse_xml()