[#2095] Rewrite of RSRGenericFilterBackend.filter_queryset()
Total rewrite using ast.literal_eval().

Using literal_eval() allows the use of literal python dictionaries as the
query string param values. So one dict is enough for all filter clauses,
and one can be specified for use in queryset.exclude() in the same
manner.

Also added the query string params select_related and prefetch related,
that take a python list literal of model names to be included in calls to
queryset.select_related() and .prefetch_related(). See the docstring for
an example.