[#2095] Legacy support for DjangoFilterBackend-based filtering
DjangoFilterBackend is supported on many of the views and is used by our
partners. Thus we can't just remove that way of being able to filter.

Fix by adding BaseRSRViewSet.get_queryset() where we look for "old style"
filters in the query string and cast the values of them so they will work
as kwargs in queryset.filter() calls.