[#2351] Fix count of mandatory fields in section 8
In section 8 there are two linked fields for the sector vocabulary and sector code. The bug where you can never fully complete section 8 is because two of the three sector code fields are always hidden, but are counted among the mandatory fields.

Fix by adding a check in markMandatoryFields() to exclude fields that have a parent with the always-hidden class.