[#2298] Merge develop and apply fixes to DRF
Update the branch with a recent version of develop and fix what's broken.

The main fix is the removal of most of the custom pagination that was added to support Tastypie in DRF 2.3. This is replaced by two new pagination classes, LimitSizePageNumberPagination and TastypieOffsetPagination as well as a rewite of BaseRSRViewSet.paginate_queryset().