[#2421] Apply escapejs filter translated strings in application/json
If a translation string has a new-line character, for instance, and it
was used inside a script tag of type `application/json`, parsing that
script tag will fail.  All translated strings in such script tags should
be filtered using `escapejs` to be considered safe.

This commit adds a test to ensure that is the case, along with fixing
all the broken templates.

Fixes #2421