[#2298] Fix failing test due to relative_data mismatch
`IndicatorPeriodData.relative_data` is a `BooleanField` with a default
set to `True`.  But, in DRF3 any forms which don't have a value set for
a `BooleanField`, the value is assumed to be `False` because HTML
checkboxes don't submit values of unchecked check boxes. To work around
this, the serializer uses a `NullBooleanField` serializer, so that the
default value of `True` from the model is made use of, when the form
doesn't submit any value for the `relative_data` field.

http://www.django-rest-framework.org/topics/release-notes/?q=BooleanField#322