[#2298] Use content_type by default unless explicitly set to None
Form submissions don't have the content_type 'application/json'.  Their
content_type is best left for the client to figure out by itself.

When `content_type` is set to 'application/json' request.body has the
data, and not request.POST.  So, we cannot be globally setting the
content_type to `application/json` for all non-xml requests.  If the
end-point expects form-data then we should take care not to set the
content type to `application/json`.

The query parameter `format` just chooses the renderer for the response.
Using that to set `content_type` is semantically INCORRECT.  This commit
sets the `content_type` to 'application/xml' when the data is a string,
because XML data is the only strings we have in our test data.