[#2424] Improve queryset filtering based on permissions
Previously, we iterated over all the private objects when filtering
querysets, to figure out which items to filter out of a queryset.  This
commit changes that code to do the filtering in the DB queries, to make
this faster.  The rules permission predicate expressions are converted
to DB queryset filters using a hackish parser.

Closes #2424