[#2398] Fix users not showing when a user has no groups
The user management page doesn't show any users when a single employment
shown on a page doesn't have a group assigned.  The post_save signal on
Employments actually tries to ensure that all Employments actually have
a Group and uses the 'Users' group as default.  But, the current DB
seems to have at least one Employment without a Group.

Fixes #2398