[#2298] Move pagination hack to settings (80-travis.conf)
Previously, settings were being patched in the `setUpClass` method of
`MigrationTestCase` and this worked when only those tests were being
run.  But, when running the whole test suite, this hack does not work.
Instead, we move this hack to settings so that we are able to overwrite
the settings correctly when the whole test suite is run.