[#2374] Set employment group default in pre_save not post_save
Employments are forced to have a non-null group, though the model does
allow null values, by trying to set a default group in the post_save.
But, this can fail if such an employment already exists, thus creating
an employment with a null group.  Multiple such employments with a null
group can get created, too.

This commit moves the setting of the group to the pre_save signal, thus
preventing an invalid employment from getting created.

Closes #2374, Closes #2122