[#2452] Add selective import of certain elements
If one or more element of the same type has the attribute akvo:import="false" all elements of that type are ignored when importing. Any existing data will remain unchanged.

The elements for which this is checked are: budget, contact-info, humanitarian-scope, legacy-data, location, participating-org, planned-disbursement, policy-marker, recipient-country, recipient-region, related-activity, result, sector and transaction.