[#2496] Accumulate and cache all the validation rules
Previously, for each field in the project editor, a DB query was made to find
out whether the field is mandatory or hidden.  This causes thousands of queries
to occur, and increases with increasing number of validation sets.  This commit
changes this code to fetch all the validation rules and cache them in memory,
while templates are being rendered.

Closes #2496