[#2447] Implement form for creating and changing Updates WIP
Components to create and modify indicator period data a.k.a. Updates.

UpdateForm is the main component, it manages the state for editing the update and talks to the backend.

NewUpdateForm is a wrapper to display the form when creating a new update.

Update renders one update using UpdateDisplay and renders UpdateForm for editing an existing update.

Header, ActualValueInput, ActualValueDescription, Attachments and UpdateFormButtons are non-stateful function based components rendering parts of UpdateForm.