[#2447] Fix problem with open updates when deleting
When deleting an update the open/closed state of other updates would change. After some head scratching I realised
that the keys generated for the Panel components in Level.render() results in the change of the key for a panel of
particular update when re-rendering after an update is deleted. This results in the Collapse activeKey prop indicating
the wrong panels as open.

Fix by assigning the model object IDs as keys for the panels, they are guaranteed to be unique and static for any
level of panels.