[#2447] Refactor Level
Turn level into an ordinary Component, rather than the base class for all the levels.

This is because I got strange behaviour (I think) in the child components around the timing of props and rendering.

Regardless this is a simpler solution that still avoids most of the code duplication that the parent-child classes were supposed to solve.