[#2447] Refactor how updates data is managed
The data for all updates is kept in App.state.models.updates, including updates being created.

This makes the handling and rendering of updates more homogenous regardless of wether it's an exiting update or a new one.

Change the NewUpdateForm component to NewUpdateButton as the same component, UpdateForm is used for editing both new and existing updates.

Refactor UpdateForm to handle editing of both new and existing updates. This results in the layout being the same for all update forms.

Add state management (state.activeKey) of the update Panels in Periods. This is needed here since NewUpdateButton is rendered here and it needs to call openNewForm().