[#2507] Fix broken pagination on subclasses of PublicProjectViewSet
Not ensuring that `.distinct` is called on the queryset used by subclasses of
`PublicProjectViewSet` caused a bug where an indicator period was missing from
the results page of a project.  This commit fixes that bug.

This commit comes with a simple test, but the test isn't actually able to
reproduce the bug on production.  Not sure what exactly needs to be done to
reproduce that bug.

Closes #2507