[#2447] Refactor Level
The Level component was getting very complicated to use. This refactor tries to simplify while still using one common component for all accordion levels.

The result is the level() function, a higher order component, i.e. a function that wraps a component (actually two for technical reasons) in common functionality and returns the result of this wrapping as a new component.

level() does two things. It manages the Collapse-Panel pairs of components for each level, and it also implements a toggle button that opens all panels of the current level.

The different levels, Results, Indicators etc, have been refactored to use level(). level() requires two components, one for the header of the Panel and one for rendering the actual level content. The reason for this is that is that I can't get Panel to work as part of a component of my own. It seems Panel has to be the direct decendant of Collapse for the accordion to work correctly and thus it has to be part of the rendering in level()