[#2540] Allow listing more projects in the directory
The query for `partnership_more_link` is very slow, and adding a
`select_related` makes it quite a lot faster.  This looks like the only query
that is slow on the project listing page.

We add a `limit` parameter to specify the number of projects to be listed in
the directory, with the default value being 10.

Closes #2540