[#2540] Add caching to codelist_value utility function
And refactor codelist_name function to use this.  These functions are called in
so many places, that caching them should bring significant benefits all over
the site.

RSR seems to already have memcached configured and it's a shame that we don't
use it anywhere except for thumbnails.  A lot of things can benefit from caching!