[#2454] Honeypot form field in registration
Add a hidden from field in the registration form. If this field contains data when a POST is made to the register view, discard the form and redirect to the RSR homepage.

The assumption is that if the honeypot field has data a bot has filled it in and we should discard the registration.