[#2447] Prevent update from being saved more than once
On a slow connection the update form stays open for a while after having pressed the save button. This can result in double posting of new updates.

Prevent double posting by disabling the save button when models.updates.changing is true, meaning that a call to the backend is underway.

Also disable the comments (notes) form for new updates, as they cannot be saved to a not yet instantiated update.