[#2447] Fix filter views opening and closing of panels
When a filtered view is applied, use ui.visibleKeys to determine if a collapsed panel shall be visible.

This way the filtered view can remain while the user can interact with the panels of the filtered view as expected.

Also includes showing new updates when they are created for a period that is visible in the filter.