[#2552] Add basic react UI with an initial data population
The data for the options are returned from a new typeahead endpoint that returns
only those options for which at there's least one project in the current view.
When filtering of projects is implemented based on the selections in the UI, the
options should automatically be filtered down to be correct.