[#2620] Update translation files for French and Spanish
Also removed the locale/all directory that must have been a mistake at some point.

Note that when running the makemessages command gettext complains about line 48 in models/sector.py.
This seems to be a bug in gettext that occurs when the translation string has two "%" somewhere in the text.
For some reason gettext thinks the second % is part of a format string.