[#2651] Don't cache groups in the code
Groups can be deleted and recreated -- this happens in tests, though this may
not happen in the real code.  This breaks tests because cached group ids are not
the same as the current ids of groups in the db.

Instead, check for permissions using group names.  Groups are fetched from the
DB using names, everywhere.  We therefore move to comparing employment group
name with canonical names for checking permissions.