[#2696] Revise update form "action buttons"
Depending on the state of an update different actions can be made. This is also dependent on the user's role.

Add an object that defines all possible states for updates and whant actions can be made depending on role.

Use the UPDATE_BUTTON object to show the correct buttons for each state/role combination.