[#2552] Add a local cache for options fetched from the server
We cache all the options fetched from the server locally, in a javascript
variable. This cache is therefore wiped out on page reloads and when filter is
applied. Though the slowest API call is fetching the options, when no filter is
selected, we cache the options for all the filter selections since clearing a
filter also sends a request and caching the previous options can make this much
faster for the user.